chore(wtr): move test configs into directory
